simple-web-app-with-microservices/user-service/requirements.txt
flask
flask-cors
requests
pyjwt[cryptography]